Fix the types for new operator
git-svn-id: https://svn.aros.org/svn/aros/trunk@34624 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/mesa/src/mesa/cpp_emul.cpp
@@ -4,8 +4,9 @@
 */
 
 #include <proto/exec.h>
+#include <stdlib.h>
 
-void * operator new(unsigned int size)
+void * operator new(size_t size)
 {
 	return AllocVec(size, MEMF_PUBLIC | MEMF_CLEAR);
 }
@@ -15,7 +16,7 @@ void operator delete(void * ptr)
 	FreeVec(ptr);
 }
 
-void * operator new[](unsigned int size)
+void * operator new[](size_t size)
 {
 	return AllocVec(size, MEMF_PUBLIC | MEMF_CLEAR);
 }
